+++ src/components/CompactPost/styles.css
