+++ src/components/PagesNavigator/styles.css
