+++ src/pages/PostDetail/components/DetailedPost/index.tsx
@@ -1,4 +1,4 @@
-import ProfileIcon from "../../../components/ProfileIcon";
+import ProfileIcon from "../../../../components/ProfileIcon";
 import "./styles.css";
 
 type PostProps = {
