+++ src/pages/PostDetail/components/DetailedPost/styles.css
