+++ src/pages/PostDetail/components/DetailedPostHeader/styles.css
