+++ src/pages/PostList/components/PostListResult/styles.css
